Added data parameter to each geom
So this changes the code significantly and was a complete pain. Still,
I’ve done it now and it passes build.
Still BiocCheck Warnings & Notes.
I’m not even sure this is a real improvement.
Still I did get rid of prot_data as a data file too which was
important.